Campaign creation works!
Campaign creation is done and adds campaigns successfully to the database
